Fix bug in Virus exploration simulation

+++ similar2logo-examples/src/main/java/fr/univ_artois/lgi2a/similar2logo/examples/virus/model/PersonFactory.java
@@ -95,7 +95,7 @@ public class PersonFactory {
 		double initialX,
 		double initialY,
 		boolean infected,
-		int timeInfected,
+		long timeInfected,
 		int lifeTime
  	){
  		if( ! category.isA(TurtleAgentCategory.CATEGORY) ) {
